feat: implements the confirmation dialog

+++ src/app/courses/containers/courses/courses.component.ts
@@ -6,6 +6,7 @@ import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/err
 import { Course } from '../../model/course';
 import { CoursesService } from '../../service/courses.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -54,17 +55,26 @@ export class CoursesComponent implements OnInit{
   }
 
   onRemove(course: Course){
-    this.coursesService.remove(course.id).subscribe(
-      () => {
-        this.refresh();
-        this.snackBar.open('Curso removido com sucesso.','X',{
-          duration: 3000,
-          verticalPosition: 'top',
-          horizontalPosition:'center'
-        });
-      },
-      error => this.onError('Erro ao tentar remover curso.')
-    )
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data:'Tem certeza que deseja remover este curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result){
+        this.coursesService.remove(course.id).subscribe(
+          () => {
+            this.refresh();
+            this.snackBar.open('Curso removido com sucesso.','X',{
+              duration: 3000,
+              verticalPosition: 'top',
+              horizontalPosition:'center'
+            });
+          },
+          error => this.onError('Erro ao tentar remover curso.')
+        );
+      }
+    });
+    
   }
 
 }
